Make parsing state width fixed to int64

+++ state.go
@@ -5,7 +5,7 @@ import "context"
 // ChangeState invokes all modification handlers fns passing them the value v.
 // The function does no affect the plate read position.
 // If any signal handler from fns returns non-nil error the function fails with that error.
-func ChangeState(v int, fns ...func(int) error) Nom {
+func ChangeState(v int64, fns ...func(int64) error) Nom {
 	return func(context.Context, Plate) (err error) {
 		for _, mod := range fns {
 			if err = mod(v); err != nil {
@@ -18,7 +18,7 @@ func ChangeState(v int, fns ...func(int) error) Nom {
 
 // RequireState runs state tests fns for against the value v. If at least one test fails the function will fail.
 // The function does no affect the plate read position.
-func RequireState(v int, fns ...func(int) bool) Nom {
+func RequireState(v int64, fns ...func(int64) bool) Nom {
 	return func(context.Context, Plate) error {
 		for _, test := range fns {
 			if !test(v) {
